Move UploadElections, and create placeholder for UploadBallots
