+++ packages/frontend/src/App.tsx
@@ -1,11 +1,10 @@
-import { useEffect } from 'react'
-import { BrowserRouter as Router, Routes, Route, useLocation } from 'react-router-dom'
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
 import { ThemeContextProvider } from './theme'
 import Header from './components/Header'
 import Election from './components/Election/Election'
 import Sandbox from './components/Sandbox'
 import LandingPage from './components/LandingPage'
-import { Box, CssBaseline, Typography } from '@mui/material'
+import { Box, CssBaseline } from '@mui/material'
 import { SnackbarContextProvider } from './components/SnackbarContext'
 import Footer from './components/Footer'
 import { ConfirmDialogProvider } from './components/ConfirmationDialogProvider'
@@ -18,7 +17,7 @@ import OpenElections from './components/Elections/OpenElections'
 import { FeatureFlagContextProvider } from './components/FeatureFlagContextProvider'
 import ComposeContextProviders from './components/ComposeContextProviders'
 import './i18n/i18n'
-import UploadElections from './components/UploadElections'
+import UploadElections from './components/Election/Upload/UploadElections'
 import Redirect from './components/Redirect'
 import PublicArchive from './components/Elections/PublicArchive'
 import NameMatchingTester from './components/NameMatchingTester'
